barking dog test

+++ src/com/test/backjun/Barkindog/Baek6198.java
@@ -11,13 +11,13 @@ public class Baek6198 {
         int n = Integer.parseInt(reader.readLine());
         int[] buildings = new int[n];
         Stack<Integer> stack = new Stack<>();
-        int cnt = 0;
+        long cnt = 0;
         for (int i = 0; i < buildings.length; i++) {
             int now = Integer.parseInt(reader.readLine());
-            while(!stack.empty() && stack.peek()<=now){
+            while (!stack.empty() && stack.peek() <= now) {
                 stack.pop();
             }
-            cnt+=stack.size();
+            cnt += stack.size();
             stack.add(now);
         }
